feat(db): very first draft of a basic persistence
This is not meant to go in production but just for benchmarking Hwedis
out so I don't care too much if things are not proper.

Still, there are a couple of things I would like to do.
